my_social_project/requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
channels==4.1.0
cmd2==2.4.3
distlib==0.3.8
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
Mypackage==0.1
mysqlclient==2.2.4
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
PyJWT==2.9.0
pyperclip==1.9.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
typing-extensions==4.12.2
virtualenv==20.26.3
wcwidth==0.2.13
